refactor: Move updateDonationPayment fn to DonationService This is where it should belong
