+++ apps/api/src/paypal/paypal.module.ts
@@ -4,9 +4,10 @@ import { PaypalService } from './paypal.service'
 import { HttpModule } from '@nestjs/axios'
 import { CampaignModule } from '../campaign/campaign.module'
 import { ConfigService } from '@nestjs/config'
+import { DonationsModule } from '../donations/donations.module'
 
 @Module({
-  imports: [HttpModule, CampaignModule],
+  imports: [HttpModule, CampaignModule, DonationsModule],
   controllers: [PaypalController],
   providers: [PaypalService, ConfigService],
   exports: [PaypalService],
