+++ apps/api/src/paypal/paypal.service.spec.ts
@@ -6,6 +6,7 @@ import { NotificationModule } from '../sockets/notifications/notification.module
 import { PaypalModule } from './paypal.module'
 import { PaypalService } from './paypal.service'
 import { MarketingNotificationsModule } from '../notifications/notifications.module'
+import { DonationsModule } from '../donations/donations.module'
 
 describe('PaypalService', () => {
   let service: PaypalService
@@ -15,6 +16,7 @@ describe('PaypalService', () => {
       imports: [
         PaypalModule,
         CampaignModule,
+        DonationsModule,
         ConfigModule,
         HttpModule,
         NotificationModule,
