+++ apps/api/src/paypal/paypal.service.ts
@@ -3,6 +3,7 @@ import { ConfigService } from '@nestjs/config'
 import { CampaignService } from '../campaign/campaign.service'
 import { HttpService } from '@nestjs/axios'
 import { PaymentStatus, DonationType, PaymentProvider, PaymentType } from '@prisma/client'
+import { DonationsService } from '../donations/donations.service'
 
 @Injectable()
 export class PaypalService {
@@ -10,6 +11,7 @@ export class PaypalService {
     private campaignService: CampaignService,
     private config: ConfigService,
     private httpService: HttpService,
+    private donationService: DonationsService,
   ) {}
 
   /*
@@ -26,7 +28,7 @@ export class PaypalService {
     // get campaign by id
     const campaign = await this.campaignService.getCampaignById(billingDetails.campaignId)
 
-    await this.campaignService.updateDonationPayment(
+    await this.donationService.updateDonationPayment(
       campaign,
       billingDetails,
       PaymentStatus.waiting,
@@ -49,7 +51,7 @@ export class PaypalService {
     // get campaign by id
     const campaign = await this.campaignService.getCampaignById(billingDetails.campaignId)
 
-    await this.campaignService.updateDonationPayment(
+    await this.donationService.updateDonationPayment(
       campaign,
       billingDetails,
       PaymentStatus.succeeded,
