+++ apps/api/src/stripe/events/stripe-payment.service.ts
@@ -20,6 +20,7 @@ import { EmailService } from '../../email/email.service'
 import { RefundDonationEmailDto } from '../../email/template.interface'
 import { PrismaService } from '../../prisma/prisma.service'
 import { StripeService } from '../stripe.service'
+import { DonationsService } from '../../donations/donations.service'
 
 /** Testing Stripe on localhost is described here:
  * https://github.com/podkrepi-bg/api/blob/master/TESTING.md#testing-stripe
@@ -31,6 +32,7 @@ export class StripePaymentService {
     private recurringDonationService: RecurringDonationService,
     private sendEmail: EmailService,
     private stripeService: StripeService,
+    private donationService: DonationsService,
   ) {}
 
   @StripeWebhookHandler('payment_intent.created')
@@ -64,7 +66,7 @@ export class StripePaymentService {
     /*
      * Handle the create event
      */
-    await this.campaignService.updateDonationPayment(campaign, paymentData, PaymentStatus.waiting)
+    await this.donationService.updateDonationPayment(campaign, paymentData, PaymentStatus.waiting)
   }
 
   @StripeWebhookHandler('payment_intent.canceled')
@@ -110,7 +112,7 @@ export class StripePaymentService {
 
     const campaign = await this.campaignService.getCampaignById(metadata.campaignId)
 
-    await this.campaignService.updateDonationPayment(campaign, billingData, PaymentStatus)
+    await this.donationService.updateDonationPayment(campaign, billingData, PaymentStatus)
   }
 
   @StripeWebhookHandler('charge.succeeded')
@@ -137,7 +139,7 @@ export class StripePaymentService {
 
     const billingData = getPaymentDataFromCharge(charge)
 
-    const donationId = await this.campaignService.updateDonationPayment(
+    const donationId = await this.donationService.updateDonationPayment(
       campaign,
       billingData,
       PaymentStatus.succeeded,
@@ -172,7 +174,7 @@ export class StripePaymentService {
 
     const campaign = await this.campaignService.getCampaignById(metadata.campaignId)
 
-    await this.campaignService.updateDonationPayment(campaign, billingData, PaymentStatus.refund)
+    await this.donationService.updateDonationPayment(campaign, billingData, PaymentStatus.refund)
 
     if (billingData.billingEmail !== undefined) {
       const recepient = { to: [billingData.billingEmail] }
@@ -365,7 +367,7 @@ export class StripePaymentService {
 
     const paymentData = getInvoiceData(invoice)
 
-    await this.campaignService.updateDonationPayment(campaign, paymentData, PaymentStatus.succeeded)
+    await this.donationService.updateDonationPayment(campaign, paymentData, PaymentStatus.succeeded)
 
     //updateDonationPayment will mark the campaign as completed if amount is reached
     await this.cancelSubscriptionsIfCompletedCampaign(metadata.campaignId)
